# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Multi-tenant monitoring SaaS platform API for website health checks, SSL monitoring, and intelligent alerting.",
        "title": "V-Insight API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/eovipmak/v-insight",
            "email": "support@v-insight.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/monitors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all monitors for the current tenant",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitors"
                ],
                "summary": "List all monitors",
                "responses": {
                    "200": {
                        "description": "List of monitors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Monitor"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new monitoring endpoint for a website or service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitors"
                ],
                "summary": "Create a new monitor",
                "parameters": [
                    {
                        "description": "Monitor configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateMonitorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Monitor created successfully",
                        "schema": {
                            "$ref": "#/definitions/entities.Monitor"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/monitors/{id}/checks": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the health check history for a specific monitor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitors"
                ],
                "summary": "Get monitor check history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of monitor checks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.MonitorCheck"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Monitor not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/monitors/{id}/ssl-status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get SSL certificate information and expiration status for a monitor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Monitors"
                ],
                "summary": "Get SSL certificate status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Monitor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSL status and alert threshold",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Monitor not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "entities.Monitor": {
            "type": "object",
            "properties": {
                "check_interval": {
                    "description": "seconds",
                    "type": "integer"
                },
                "check_ssl": {
                    "description": "whether to check SSL certificate",
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "last_checked_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "ssl_alert_days": {
                    "description": "days before expiry to alert",
                    "type": "integer"
                },
                "tenant_id": {
                    "type": "integer"
                },
                "timeout": {
                    "description": "seconds",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "entities.MonitorCheck": {
            "type": "object",
            "properties": {
                "checked_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "monitor_id": {
                    "type": "string"
                },
                "response_time_ms": {
                    "type": "integer"
                },
                "ssl_expires_at": {
                    "type": "string"
                },
                "ssl_valid": {
                    "type": "boolean"
                },
                "status_code": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.CreateMonitorRequest": {
            "type": "object",
            "required": [
                "name",
                "url"
            ],
            "properties": {
                "check_interval": {
                    "description": "minimum 60 seconds",
                    "type": "integer",
                    "minimum": 60
                },
                "check_ssl": {
                    "description": "pointer to allow explicit false",
                    "type": "boolean"
                },
                "enabled": {
                    "description": "pointer to allow explicit false",
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "ssl_alert_days": {
                    "description": "minimum 1 day",
                    "type": "integer",
                    "minimum": 1
                },
                "timeout": {
                    "description": "5-120 seconds",
                    "type": "integer",
                    "maximum": 120,
                    "minimum": 5
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}